Restructure files in plugin pkg.
